Utilities

The two main general utilities for working with quantities are ustrip and dimension:

DynamicQuantities.ustrip — Function

ustrip(q::AbstractQuantity)

Remove the units from a quantity.

DynamicQuantities.dimension — Function

dimension(q::AbstractQuantity)

Get the dimensions of a quantity, returning an AbstractDimensions object.

Accessing dimensions

Utility functions to extract specific dimensions are as follows:

DynamicQuantities.ulength — Function

ulength(q::AbstractQuantity)
ulength(d::AbstractDimensions)

Get the length dimension of a quantity (e.g., meters^(ulength)).

DynamicQuantities.umass — Function

umass(q::AbstractQuantity)
umass(d::AbstractDimensions)

Get the mass dimension of a quantity (e.g., kg^(umass)).

DynamicQuantities.utime — Function

utime(q::AbstractQuantity)
utime(d::AbstractDimensions)

Get the time dimension of a quantity (e.g., s^(utime))

DynamicQuantities.ucurrent — Function

ucurrent(q::AbstractQuantity)
ucurrent(d::AbstractDimensions)

Get the current dimension of a quantity (e.g., A^(ucurrent)).

DynamicQuantities.utemperature — Function

utemperature(q::AbstractQuantity)
utemperature(d::AbstractDimensions)

Get the temperature dimension of a quantity (e.g., K^(utemperature)).

DynamicQuantities.uluminosity — Function

uluminosity(q::AbstractQuantity)
uluminosity(d::AbstractDimensions)

Get the luminosity dimension of a quantity (e.g., cd^(uluminosity)).

DynamicQuantities.uamount — Function

uamount(q::AbstractQuantity)
uamount(d::AbstractDimensions)

Get the amount dimension of a quantity (e.g., mol^(uamount)).

Internals

FixedRational

DynamicQuantities.FixedRational — Type

FixedRational{T,den}

A rational number with a fixed denominator. Significantly faster than Rational{T}, as it never needs to compute the gcd apart from when printing. Access the denominator with denom(F) (which converts to T).

Fields

num: numerator of type T. The denominator is fixed to the type parameter den.

DynamicQuantities.denom — Function

denom(F::FixedRational)

Since den can be a different type than T, this function is used to get the denominator as a T.