Symbolic Dimensions

Whereas u"..." will automatically convert all units to the same base SI units, us"..." will not. This uses the SymbolicDimensions type, which is a subtype of AbstractDimensions that stores the dimensions symbolically. This is useful for keeping track of the original units and constants in a user-entered expression.

The two main functions for working with symbolic units are sym_uparse and us_str:

DynamicQuantities.@us_str — Macro

us"[unit expression]"

Parse a string containing an expression of units and return the corresponding Quantity object with Float64 value. However, unlike the regular u"..." macro, this macro uses SymbolicDimensions for the dimension type, which means that all units and constants are stored symbolically and will not automatically expand to SI units. For example, us"km/s^2" would be parsed to Quantity(1.0, SymbolicDimensions, km=1, s=-2).

Note that inside this expression, you also have access to the Constants module. So, for example, us"Constants.c^2 * Hz^2" would evaluate to Quantity(1.0, SymbolicDimensions, c=2, Hz=2). However, note that due to namespace collisions, a few physical constants are automatically converted.

DynamicQuantities.SymbolicUnitsParse.sym_uparse — Function

sym_uparse(raw_string::AbstractString)

Parse a string containing an expression of units and return the corresponding Quantity object with Float64 value. However, that unlike the regular u"..." macro, this macro uses SymbolicDimensions for the dimension type, which means that all units and constants are stored symbolically and will not automatically expand to SI units. For example, sym_uparse("km/s^2") would be parsed to Quantity(1.0, SymbolicDimensions, km=1, s=-2).

Note that inside this expression, you also have access to the Constants module. So, for example, sym_uparse("Constants.c^2 * Hz^2") would evaluate to Quantity(1.0, SymbolicDimensions, c=2, Hz=2). However, note that due to namespace collisions, a few physical constants are automatically converted.

To convert a quantity to its regular base SI units, use expand_units:

DynamicQuantities.expand_units — Function

expand_units(q::Quantity{<:Any,<:SymbolicDimensions})

Expand the symbolic units in a quantity to their base SI form. In other words, this converts a Quantity with SymbolicDimensions to one with Dimensions.