Types

DynamicQuantities.Quantity — Type

Quantity{T,D<:AbstractDimensions} <: AbstractQuantity{T,D}

Physical quantity with value value of type T and dimensions dimensions of type D. For example, the velocity of an object with mass 1 kg and velocity 2 m/s is Quantity(2, mass=1, length=1, time=-1). You should access these fields with ustrip(q), and dimension(q). You can access specific dimensions with ulength(q), umass(q), utime(q), ucurrent(q), utemperature(q), uluminosity(q), and uamount(q).

Severals operators in Base are extended to work with Quantity objects, including *, +, -, /, abs, ^, sqrt, and cbrt, which manipulate dimensions according to the operation.

Fields

value::T: value of the quantity of some type T. Access with ustrip(::Quantity)

dimensions::D: dimensions of the quantity. Access with dimension(::Quantity)

Constructors

Quantity(x; kws...): Construct a quantity with value x and dimensions given by the keyword arguments. The value type is inferred from x. R is set to DEFAULT_DIM_TYPE.

Quantity(x, ::Type{D}; kws...): Construct a quantity with value x with dimensions given by the keyword arguments, and the dimensions type set to D.

Quantity(x, d::D): Construct a quantity with value x and dimensions d of type D.

Quantity{T}(...): As above, but converting the value to type T. You may also pass a Quantity as input.

Quantity{T,D}(...): As above, but converting the value to type T and dimensions to D. You may also pass a Quantity as input.

DynamicQuantities.Dimensions — Type

Dimensions{R<:Real} <: AbstractDimensions{R}

A type representing the dimensions of a quantity, with each field giving the power of the corresponding dimension. For example, the dimensions of velocity are Dimensions(length=1, time=-1). Each of the 7 dimensions are stored using the type R, which is by default a rational number.

Fields

length: length dimension (i.e., meters^(length))

mass: mass dimension (i.e., kg^(mass))

time: time dimension (i.e., s^(time))

current: current dimension (i.e., A^(current))

temperature: temperature dimension (i.e., K^(temperature))

luminosity: luminosity dimension (i.e., cd^(luminosity))

amount: amount dimension (i.e., mol^(amount))

Constructors

Dimensions(args...): Pass all the dimensions as arguments. R is set to DEFAULT_DIM_BASE_TYPE.

Dimensions(; kws...): Pass a subset of dimensions as keyword arguments. R is set to DEFAULT_DIM_BASE_TYPE.

Dimensions(::Type{R}; kws...) or Dimensions{R}(; kws...): Pass a subset of dimensions as keyword arguments, with the output type set to Dimensions{R}.

Dimensions{R}(): Create a dimensionless object typed as Dimensions{R}.

Dimensions{R}(d::Dimensions): Copy the dimensions from another Dimensions object, with the output type set to Dimensions{R}.

There are also abstract types available. There are no required functions to build an interface, most relevant functions are defined on the abstract functions (including constructors).

DynamicQuantities.AbstractDimensions — Type

AbstractDimensions{R}

An abstract type for dimension types. R is the type of the exponents of the dimensions, and by default is set to DynamicQuantities.DEFAULT_DIM_BASE_TYPE. AbstractDimensions are used to store the dimensions of AbstractQuantity objects. Together these enable many operators in Base to manipulate dimensions. This type has generic constructors for creating dimension objects, so user-defined dimension types can be created by simply subtyping AbstractDimensions, without the need to define many other functions.

The key function that one could wish to overload is DynamicQuantities.dimension_name(::AbstractDimensions, k::Symbol) for mapping from a field name to a base unit (e.g., length by default maps to m). You may also need to overload DynamicQuantities.constructor_of(::Type{T}) in case of non-standard construction.

DynamicQuantities.AbstractQuantity — Type

AbstractQuantity{T,D}

An abstract type for quantities. T is the type of the value of the quantity, and D is the type of the dimensions of the quantity. By default, D is set to DynamicQuantities.DEFAULT_DIM_TYPE. T is inferred from the value in a calculation, but in other cases is defaulted to DynamicQuantities.DEFAULT_VALUE_TYPE. It is assumed that the value is stored in the :value field, and the dimensions object is stored in the :dimensions field. These fields can be accessed with ustrip and dimension, respectively. Many operators in Base are defined on AbstractQuantity objects, including +, -, *, /, ^, sqrt, cbrt, abs.

Note also that the Quantity object can take a custom AbstractDimensions as input, so there is often no need to subtype AbstractQuantity separately.

Another type which subtypes AbstractDimensions is SymbolicDimensions:

DynamicQuantities.SymbolicDimensions — Type

SymbolicDimensions{R} <: AbstractDimensions{R}

An AbstractDimensions with one dimension for every unit and constant symbol. This is to allow for lazily reducing to SI base units, whereas Dimensions is always in SI base units. Furthermore, SymbolicDimensions stores dimensions using a sparse vector for efficiency (since there are so many unit symbols).

You can convert a quantity using SymbolicDimensions as its dimensions to one which uses Dimensions as its dimensions (i.e., base SI units) expand_units.