Units

The two main functions for working with units are uparse and u_str:

DynamicQuantities.UnitsParse.@u_str — Macro

u"[unit expression]"

Parse a string containing an expression of units and return the corresponding Quantity object with Float64 value. For example, u"km/s^2" would be parsed to Quantity(1000.0, length=1, time=-2).

Note that inside this expression, you also have access to the Constants module. So, for example, u"Constants.c^2 * Hz^2" would evaluate to the quantity corresponding to the speed of light multiplied by Hertz, squared.

DynamicQuantities.UnitsParse.uparse — Function

uparse(s::AbstractString)

Parse a string containing an expression of units and return the corresponding Quantity object with Float64 value. For example, uparse("m/s") would be parsed to Quantity(1.0, length=1, time=-1).

Note that inside this expression, you also have access to the Constants module. So, for example, uparse("Constants.c^2 * Hz^2") would evaluate to the quantity corresponding to the speed of light multiplied by Hertz, squared.

Available units

The base SI units are as follows. Instead of calling directly, it is recommended to access them via the @u_str macro, which evaluates the expression in a namespace with all the units available.

DynamicQuantities.Units.m — Constant

Length in meters. Available variants: fm, pm, nm, μm (/um), cm, inch, dm, mm, ft, km, mi, Mm, Gm.

DynamicQuantities.Units.kg — Constant

Mass in kilograms. Available variants: pg, ng, μg (/ug), mg, g.

DynamicQuantities.Units.s — Constant

Time in seconds. Available variants: fs, ps, ns, μs (/us), ms, min (/minute), h (/hr), day (/d), wk, yr, kyr, Myr, Gyr.

DynamicQuantities.Units.A — Constant

Current in Amperes. Available variants: nA, μA (/uA), mA, kA.

DynamicQuantities.Units.K — Constant

Temperature in Kelvin. Available variant: mK.

DynamicQuantities.Units.cd — Constant

Luminosity in candela. Available variant: mcd.

DynamicQuantities.Units.mol — Constant

Amount in moles. Available variant: pmol, nmol, μmol (/umol), mmol.

Derived units

Several derived SI units are available as well:

DynamicQuantities.Units.Hz — Constant

Frequency in Hertz. Available variants: nHz, μHz (/uHz), mHz, kHz, MHz, GHz.

DynamicQuantities.Units.N — Constant

Force in Newtons.

DynamicQuantities.Units.Pa — Constant

Pressure in Pascals. Available variant: kPa.

DynamicQuantities.Units.J — Constant

Energy in Joules. Available variant: kJ.

DynamicQuantities.Units.W — Constant

Power in Watts. Available variants: mW, kW, MW, GW.

DynamicQuantities.Units.C — Constant

Charge in Coulombs.

DynamicQuantities.Units.V — Constant

Voltage in Volts. Available variants: pV, nV, μV (/uV), mV, kV,MV,GV`.

DynamicQuantities.Units.F — Constant

Capacitance in Farads. Available variants: fF, pF, nF, μF (/uF), mF.

DynamicQuantities.Units.Ω — Constant

Resistance in Ohms. Available variant: nΩ, μΩ (/uΩ), mΩ, kΩ, MΩ, GΩ. Also available is ASCII ohm (with variants nohm, μohm (/uohm), mohm, kohm, Mohm, Gohm).

DynamicQuantities.Units.T — Constant

Magnetic flux density in Teslas.

DynamicQuantities.Units.L — Constant

Volume in liters. Available variants: μL (/uL), mL, cL, dL.

DynamicQuantities.Units.bar — Constant

Pressure in bars. Available variants: mbar.