Usage

Types

DynamicQuantities.Quantity — Type

Quantity{T,D}

Physical quantity with value value of type T and dimensions dimensions of type D. For example, the velocity of an object with mass 1 kg and velocity 2 m/s is Quantity(2, mass=1, length=1, time=-1). You should access these fields with ustrip(q), and dimensions(q). You can access specific dimensions with ulength(q), umass(q), utime(q), ucurrent(q), utemperature(q), uluminosity(q), and uamount(q).

Severals operators in Base are extended to work with Quantity objects, including *, +, -, /, abs, ^, sqrt, and cbrt, which manipulate dimensions according to the operation.

Fields

value::T: value of the quantity of some type T. Access with ustrip(::Quantity)

dimensions::D: dimensions of the quantity. Access with dimension(::Quantity)

Constructors

Quantity(x; kws...): Construct a quantity with value x and dimensions given by the keyword arguments. The value type is inferred from x. R is set to DEFAULT_DIM_TYPE.

Quantity(x, ::Type{D}; kws...): Construct a quantity with value x with dimensions given by the keyword arguments, and the dimensions type set to D.

Quantity(x, d::D): Construct a quantity with value x and dimensions d of type D.

Quantity{T}(...): As above, but converting the value to type T. You may also pass a Quantity as input.

Quantity{T,D}(...): As above, but converting the value to type T and dimensions to D. You may also pass a Quantity as input.

DynamicQuantities.Dimensions — Type

Dimensions{R}

A type representing the dimensions of a quantity, with each field giving the power of the corresponding dimension. For example, the dimensions of velocity are Dimensions(length=1, time=-1). Each of the 7 dimensions are stored using the type R, which is by default a rational number.

Fields

length: length dimension (i.e., meters^(length))

mass: mass dimension (i.e., kg^(mass))

time: time dimension (i.e., s^(time))

current: current dimension (i.e., A^(current))

temperature: temperature dimension (i.e., K^(temperature))

luminosity: luminosity dimension (i.e., cd^(luminosity))

amount: amount dimension (i.e., mol^(amount))

Constructors

Dimensions(args...): Pass all the dimensions as arguments. R is set to DEFAULT_DIM_BASE_TYPE.

Dimensions(; kws...): Pass a subset of dimensions as keyword arguments. R is set to DEFAULT_DIM_BASE_TYPE.

Dimensions(::Type{R}; kws...) or Dimensions{R}(; kws...): Pass a subset of dimensions as keyword arguments, with the output type set to Dimensions{R}.

Dimensions{R}(): Create a dimensionless object typed as Dimensions{R}.

Dimensions{R}(d::Dimensions): Copy the dimensions from another Dimensions object, with the output type set to Dimensions{R}.

Utilities

The two main general utilities for working with quantities are ustrip and dimension:

DynamicQuantities.ustrip — Function

ustrip(q::AbstractQuantity)

Remove the units from a quantity.

DynamicQuantities.dimension — Function

dimension(q::AbstractQuantity)

Get the dimensions of a quantity, returning an AbstractDimensions object.

Accessing dimensions

Utility functions to extract specific dimensions are as follows:

DynamicQuantities.ulength — Function

ulength(q::AbstractQuantity)
ulength(d::AbstractDimensions)

Get the length dimension of a quantity (e.g., meters^(ulength)).

DynamicQuantities.umass — Function

umass(q::AbstractQuantity)
umass(d::AbstractDimensions)

Get the mass dimension of a quantity (e.g., kg^(umass)).

DynamicQuantities.utime — Function

utime(q::AbstractQuantity)
utime(d::AbstractDimensions)

Get the time dimension of a quantity (e.g., s^(utime))

DynamicQuantities.ucurrent — Function

ucurrent(q::AbstractQuantity)
ucurrent(d::AbstractDimensions)

Get the current dimension of a quantity (e.g., A^(ucurrent)).

DynamicQuantities.utemperature — Function

utemperature(q::AbstractQuantity)
utemperature(d::AbstractDimensions)

Get the temperature dimension of a quantity (e.g., K^(utemperature)).

DynamicQuantities.uluminosity — Function

uluminosity(q::AbstractQuantity)
uluminosity(d::AbstractDimensions)

Get the luminosity dimension of a quantity (e.g., cd^(uluminosity)).

DynamicQuantities.uamount — Function

uamount(q::AbstractQuantity)
uamount(d::AbstractDimensions)

Get the amount dimension of a quantity (e.g., mol^(uamount)).

Units

The two main functions for working with units are uparse and u_str:

DynamicQuantities.Units.@u_str — Macro

u"[unit expression]"

Parse a string containing an expression of units and return the corresponding Quantity object with Float64 value. For example, u"km/s^2" would be parsed to Quantity(1000.0, length=1, time=-2).

DynamicQuantities.Units.uparse — Function

uparse(s::AbstractString)

Parse a string containing an expression of units and return the corresponding Quantity object with Float64 value. For example, uparse("m/s") would be parsed to Quantity(1.0, length=1, time=-1).

Available units

The base SI units are as follows. Instead of calling directly, it is recommended to access them via the @u_str macro, which evaluates the expression in a namespace with all the units available.

DynamicQuantities.Units.m — Constant

Length in meters. Available variants: fm, pm, nm, μm (/um), cm, dm, mm, km, Mm, Gm.

DynamicQuantities.Units.g — Constant

Mass in grams. Available variants: μg (/ug), mg, kg.

DynamicQuantities.Units.s — Constant

Time in seconds. Available variants: fs, ps, ns, μs (/us), ms, min, h (/hr), day, yr, kyr, Myr, Gyr.

DynamicQuantities.Units.A — Constant

Current in Amperes. Available variants: nA, μA (/uA), mA, kA.

DynamicQuantities.Units.K — Constant

Temperature in Kelvin. Available variant: mK.

DynamicQuantities.Units.cd — Constant

Luminosity in candela. Available variant: mcd.

DynamicQuantities.Units.mol — Constant

Amount in moles. Available variant: mmol.

Several derived SI units are available as well:

DynamicQuantities.Units.Hz — Constant

Frequency in Hertz. Available variants: kHz, MHz, GHz.

DynamicQuantities.Units.N — Constant

Force in Newtons.

DynamicQuantities.Units.Pa — Constant

Pressure in Pascals. Available variant: kPa.

DynamicQuantities.Units.J — Constant

Energy in Joules. Available variant: kJ.

DynamicQuantities.Units.W — Constant

Power in Watts. Available variants: kW, MW, GW.

DynamicQuantities.Units.C — Constant

Charge in Coulombs.

DynamicQuantities.Units.V — Constant

Voltage in Volts. Available variants: kV, MV, GV.

DynamicQuantities.Units.F — Constant

Capacitance in Farads.

DynamicQuantities.Units.Ω — Constant

Resistance in Ohms. Available variant: mΩ. Also available is ASCII ohm (with variant mohm).

DynamicQuantities.Units.T — Constant

Magnetic flux density in Teslas.

DynamicQuantities.Units.L — Constant

Volume in liters. Available variants: mL, dL.

DynamicQuantities.Units.bar — Constant

Pressure in bars.

Internals

FixedRational

DynamicQuantities.FixedRational — Type

FixedRational{T,den}

A rational number with a fixed denominator. Significantly faster than Rational{T}, as it never needs to compute the gcd apart from when printing. Access the denominator with denom(F) (which converts to T).

Fields

num: numerator of type T. The denominator is fixed to the type parameter den.

DynamicQuantities.denom — Function

denom(F::FixedRational)

Since den can be a different type than T, this function is used to get the denominator as a T.