API Reference

API Reference for DynamicQuantities.

DynamicQuantities.Dimensions

DynamicQuantities.Quantity

DynamicQuantities.dimension

DynamicQuantities.uamount

DynamicQuantities.ucurrent

DynamicQuantities.ulength

DynamicQuantities.uluminosity

DynamicQuantities.umass

DynamicQuantities.ustrip

DynamicQuantities.utemperature

DynamicQuantities.utime

DynamicQuantities.Dimensions — Type

Dimensions

A type representing the dimensions of a quantity, with each field giving the power of the corresponding dimension. For example, the dimensions of velocity are Dimensions(length=1, time=-1).

Fields

length: length dimension (i.e., meters^(length))

mass: mass dimension (i.e., kg^(mass))

time: time dimension (i.e., s^(time))

current: current dimension (i.e., A^(current))

temperature: temperature dimension (i.e., K^(temperature))

luminosity: luminosity dimension (i.e., cd^(luminosity))

amount: amount dimension (i.e., mol^(amount))

DynamicQuantities.Quantity — Type

Quantity{T}

Physical quantity with value value of type T and dimensions dimensions. For example, the velocity of an object with mass 1 kg and velocity 2 m/s is Quantity(2, mass=1, length=1, time=-1). You should access these fields with ustrip(q), and dimensions(q). You can access specific dimensions with ulength(q), umass(q), utime(q), ucurrent(q), utemperature(q), uluminosity(q), and uamount(q).

Severals operators in Base are extended to work with Quantity objects, including *, +, -, /, ^, sqrt, and cbrt.

Fields

value::T: value of the quantity of some type T

dimensions::Dimensions: dimensions of the quantity

DynamicQuantities.dimension — Method

dimension(q::Quantity)

Get the dimensions of a quantity, returning a Dimensions object.

DynamicQuantities.uamount — Method

uamount(q::Quantity)

Get the amount dimension of a quantity (e.g., mol^(uamount)).

DynamicQuantities.ucurrent — Method

ucurrent(q::Quantity)

Get the current dimension of a quantity (e.g., A^(ucurrent)).

DynamicQuantities.ulength — Method

ulength(q::Quantity)

Get the length dimension of a quantity (e.g., meters^(ulength)).

DynamicQuantities.uluminosity — Method

uluminosity(q::Quantity)

Get the luminosity dimension of a quantity (e.g., cd^(uluminosity)).

DynamicQuantities.umass — Method

umass(q::Quantity)

Get the mass dimension of a quantity (e.g., kg^(umass)).

DynamicQuantities.ustrip — Method

ustrip(q::Quantity)

Remove the units from a quantity.

DynamicQuantities.utemperature — Method

utemperature(q::Quantity)

Get the temperature dimension of a quantity (e.g., K^(utemperature)).

DynamicQuantities.utime — Method

utime(q::Quantity)

Get the time dimension of a quantity (e.g., s^(utime))